+++ packages/admin-api/src/clients/JournalClient.ts
@@ -1,4 +1,4 @@
-import { HttpClient } from '@halo-dev/rest-api-client'
+import { HttpClient } from '@nineya/halo-rest-api-client'
 import { buildPath } from '../url'
 import { Journal, JournalParam, JournalQuery, JournalWithCmtCount, Page, Response } from '../types'