+++ packages/content-api/src/ContentApiClient.ts
@@ -1,4 +1,4 @@
-import { HaloRestAPIClient, HttpClient } from '@halo-dev/rest-api-client'
+import { HaloRestAPIClient, HttpClient } from '@nineya/halo-rest-api-client'
 import {
   ArchiveClient,
   CategoryClient,