+++ packages/content-api/src/clients/SheetClient.ts
@@ -1,4 +1,4 @@
-import { HttpClient } from '@halo-dev/rest-api-client'
+import { HttpClient } from '@nineya/halo-rest-api-client'
 import { buildPath } from '../url'
 import { ContentQuery, Page, PageQuery, Response, SheetDetail, SheetList } from '../types'