+++ packages/content-api/src/index.ts
@@ -1,4 +1,4 @@
 export { ContentApiClient } from './ContentApiClient'
-export * from '@halo-dev/rest-api-client'
+export * from '@nineya/halo-rest-api-client'
 export * from './types'
 export * from './clients'