+++ packages/content-api/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @halo-dev/content-api
+# @nineya/halo-content-api
 
 ## 1.1.0
 
@@ -9,10 +9,10 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @halo-dev/rest-api-client@1.1.0
+  - @nineya/halo-rest-api-client@1.1.0
 
 ## 1.0.0
 
 ### Patch Changes
 
-- @halo-dev/rest-api-client@1.0.0
+- @nineya/halo-rest-api-client@1.0.0